+++ package.json
@@ -14,6 +14,7 @@
     "@babel/core": "^7.10.2",
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-react": "^7.10.1",
+    "@jest/globals": "^26.0.1",
     "babel-jest": "26.0.1",
     "babel-loader": "^8.1.0",
     "enzyme": "3.11.0",
@@ -31,7 +32,6 @@
     "node": "10"
   },
   "dependencies": {
-    "@jest/globals": "^26.0.1",
     "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1"
