+++ src/components/app/app.test.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import {describe, expect, it} from '@jest/globals';
 import {App} from './app.jsx';
 
 const testData = {
